+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @memo28/utils
 
+## 1.2.3
+
+### Patch Changes
+
+- 修复utils readingWritingSeparationUtilsType 函数生成get set 函数首字母大小写问题，新增open-source-tool，enhance-redux package
+
 ## 1.2.2
 
 ### Patch Changes